Module: Overlay wrapper into place instead of adding it manually
This patch makes the module overlay the flatpak binary with the
wrapper instead of manually adding it to systemd packages, etc.

This makes enabling the module compatible with
services.flatpak.enable = true;

This patch also removes the package module option as its not really
compatible with the overlay approach and makes the the flatpak-pkg
input to the function in pkg.nix actually used.